requirements/local.txt
-r testing.txt

Sphinx

django-extensions==1.7.8

django-debug-toolbar==1.7

